split off config, pipeline, and features from database.go
