+++ pkg/database/database.go
@@ -9,8 +9,6 @@ import (
 	"fmt"
 	"log"
 	"math"
-	"net"
-	"net/url"
 	"os"
 	"path/filepath"
 	"runtime"
@@ -31,356 +29,6 @@ type Database struct {
 	duckCleanup func() error        // Cleanup hook that removes the temp directory when the process shuts down
 }
 
-// Close releases the shared connection and cleans up per-engine scratch space. We keep
-// the logic simple and sequential so shutdown paths remain predictable, matching the Go
-// Proverb "Clear is better than clever" while still returning the first encountered error.
-func (db *Database) Close() error {
-	if db == nil {
-		return nil
-	}
-
-	var firstErr error
-	if db.DB != nil {
-		if err := db.DB.Close(); err != nil && firstErr == nil {
-			firstErr = err
-		}
-	}
-
-	if db.duckCleanup != nil {
-		if err := db.duckCleanup(); err != nil && firstErr == nil {
-			firstErr = err
-		}
-	}
-
-	return firstErr
-}
-
-// serializedJob represents a unit of work that must run in isolation for engines such as
-// DuckDB, SQLite, and Chai. We avoid mutexes and keep coordination explicit with channels,
-// following the Go Proverb "Don't communicate by sharing memory; share memory by
-// communicating." Each job carries its own response channel so callers can await results
-// without blocking the worker loop.
-type serializedJob struct {
-	ctx    context.Context
-	fn     func(context.Context, *sql.DB) (any, error)
-	result chan serializedResult
-	kind   WorkloadKind
-}
-
-// serializedResult holds the outcome of a serializedJob. Keeping it separate helps the
-// worker return values without coupling callers to concrete types.
-type serializedResult struct {
-	value any
-	err   error
-}
-
-// WorkloadKind enumerates the separate queues we run through select/case so realtime feeds,
-// archive imports, user uploads, and web readers time-share the single connection without
-// letting one long backlog starve the others. Using an explicit type keeps the routing logic
-// readable and mirrors the Go Proverb "Make the zero value useful" by defaulting to the
-// general queue.
-type WorkloadKind int
-
-const (
-	WorkloadGeneral    WorkloadKind = iota // catch-all when callers do not care
-	WorkloadWebRead                        // API/UI fetches for map tiles and history
-	WorkloadUserUpload                     // single track uploads from the web form
-	WorkloadArchive                        // large TGZ imports and bulk archive loaders
-	WorkloadRealtime                       // live Safecast device updates
-)
-
-// serializedPipeline owns per-workload channels and orchestrates fair selection so one busy
-// feed cannot monopolize the single-writer engines. We avoid mutexes and keep the rotation
-// explicit so maintainers can reason about how different callers interleave.
-type serializedPipeline struct {
-	lanes []chan serializedJob // Indexed by workloadKind so selection happens without switches
-	order []int                // Round-robin order to rotate through lanes fairly
-	turn  int                  // Tracks the next lane index to probe first
-}
-
-// duckDBMaintenance serialises maintenance tasks like CHECKPOINT and VACUUM so the
-// main pipeline stays free for imports and readers. A tiny job queue and a single
-// worker goroutine keep the coordination straightforward while avoiding mutexes.
-// We scope this to DuckDB because other engines have their own housekeeping.
-type duckDBMaintenance struct {
-	db   *Database
-	jobs chan duckDBMaintenanceJob
-}
-
-// duckDBMaintenanceJob wraps a maintenance request with its logging and completion
-// channel. Keeping it explicit makes the worker loop easy to follow and lets the
-// caller await completion without blocking the queue.
-type duckDBMaintenanceJob struct {
-	ctx  context.Context
-	logf func(string, ...any)
-	done chan error
-}
-
-// serializedWaitFloor keeps serialized operations from failing fast when long-running archive jobs occupy the
-// queue. We deliberately stretch deadlines so single-writer engines like Chai, SQLite, and DuckDB can finish
-// the work instead of racing short timeouts. The helper below centralises the policy so callers stay small.
-const serializedWaitFloor = 45 * time.Second
-
-// startSerializedPipeline spins up a goroutine that owns the shared *sql.DB for single-user
-// engines. The worker picks jobs in a round-robin order across workload queues so realtime
-// bursts or TGZ imports cannot starve web readers. We stay within channels/select and avoid
-// mutexes to keep the coordination simple and observable.
-func startSerializedPipeline(db *sql.DB) *serializedPipeline {
-	// We allocate lanes in the same index order as workloadKind values so callers can
-	// select the lane via array indexing instead of switches. This keeps channel usage
-	// explicit and follows the Go proverb "A little copying is better than a little
-	// dependency" by avoiding extra indirection.
-	p := &serializedPipeline{
-		lanes: []chan serializedJob{
-			make(chan serializedJob, 64),  // WorkloadGeneral
-			make(chan serializedJob, 128), // WorkloadWebRead
-			make(chan serializedJob, 32),  // WorkloadUserUpload
-			make(chan serializedJob, 32),  // WorkloadArchive
-			make(chan serializedJob, 128), // WorkloadRealtime
-		},
-		order: []int{int(WorkloadRealtime), int(WorkloadArchive), int(WorkloadUserUpload), int(WorkloadWebRead), int(WorkloadGeneral)},
-	}
-
-	// getNextJob rotates over queues with a non-blocking pass before blocking so we keep
-	// latency low for lighter queues even while a heavy import continues to stream.
-	getNextJob := func() serializedJob {
-		// First attempt: non-blocking sweep in round-robin order.
-		for i := 0; i < len(p.order); i++ {
-			laneIdx := p.order[p.turn%len(p.order)]
-			p.turn++
-			select {
-			case job := <-p.lanes[laneIdx]:
-				return job
-			default:
-			}
-		}
-
-		// Second attempt: block until any queue receives work to avoid busy waiting.
-		select {
-		case job := <-p.lanes[p.order[0]]:
-			return job
-		case job := <-p.lanes[p.order[1]]:
-			return job
-		case job := <-p.lanes[p.order[2]]:
-			return job
-		case job := <-p.lanes[p.order[3]]:
-			return job
-		case job := <-p.lanes[p.order[4]]:
-			return job
-		}
-	}
-
-	go func() {
-		defer func() {
-			// keep the pattern explicit: ownership stays with the creator, not the worker
-		}()
-
-		for {
-			job := getNextJob()
-
-			select {
-			case <-job.ctx.Done():
-				job.result <- serializedResult{nil, job.ctx.Err()}
-				continue
-			default:
-			}
-
-			value, err := job.fn(job.ctx, db)
-			job.result <- serializedResult{value, err}
-		}
-	}()
-
-	return p
-}
-
-// startDuckDBMaintenance spins a goroutine that runs maintenance tasks one at a time so
-// they never overlap with each other. We keep the queue small to avoid unbounded memory
-// growth and rely on the serialized database pipeline to keep CHECKPOINT/VACUUM from
-// racing regular queries.
-func startDuckDBMaintenance(db *Database) *duckDBMaintenance {
-	m := &duckDBMaintenance{db: db, jobs: make(chan duckDBMaintenanceJob, 4)}
-
-	go func() {
-		for job := range m.jobs {
-			logf := job.logf
-			if logf == nil {
-				logf = log.Printf
-			}
-
-			if job.ctx == nil {
-				job.ctx = context.Background()
-			}
-
-			select {
-			case <-job.ctx.Done():
-				job.done <- job.ctx.Err()
-				close(job.done)
-				continue
-			default:
-			}
-
-			// Give the maintenance enough time to finish on large files while still respecting caller
-			// cancellations. We wrap it with queueFriendlyContext to keep the serialized worker from
-			// dropping jobs too aggressively under load.
-			maintenanceCtx, cancel := queueFriendlyContext(job.ctx, 90*time.Minute)
-			err := m.db.withSerializedConnectionFor(maintenanceCtx, WorkloadArchive, func(runCtx context.Context, conn *sql.DB) error {
-				return runDuckDBMaintenance(runCtx, conn, logf)
-			})
-			cancel()
-
-			select {
-			case job.done <- err:
-			default:
-			}
-			close(job.done)
-		}
-	}()
-
-	return m
-}
-
-// enqueue places a maintenance job onto the queue without blocking the caller. We rely on
-// select/case to keep the loop responsive to cancellation while still guaranteeing that the
-// job either enters the queue or reports the context error.
-func (m *duckDBMaintenance) enqueue(ctx context.Context, logf func(string, ...any)) <-chan error {
-	if m == nil {
-		done := make(chan error, 1)
-		close(done)
-		return done
-	}
-
-	job := duckDBMaintenanceJob{ctx: ctx, logf: logf, done: make(chan error, 1)}
-
-	for {
-		select {
-		case <-ctx.Done():
-			job.done <- ctx.Err()
-			close(job.done)
-			return job.done
-		case m.jobs <- job:
-			return job.done
-		default:
-			runtime.Gosched()
-		}
-	}
-}
-
-// scheduleDuckDBMaintenance exposes a channel-driven hook for callers that want to run a
-// checkpoint/optimize/vacuum cycle after heavy imports. We only wire it for the DuckDB
-// driver because other engines handle maintenance differently.
-func (db *Database) ScheduleDuckDBMaintenance(ctx context.Context, logf func(string, ...any)) <-chan error {
-	if db == nil || db.Driver != "duckdb" || db.upkeep == nil {
-		done := make(chan error, 1)
-		close(done)
-		return done
-	}
-	if ctx == nil {
-		ctx = context.Background()
-	}
-	return db.upkeep.enqueue(ctx, logf)
-}
-
-// serializedEnabled reports whether the database should route operations through the
-// single-owner pipeline.
-func (db *Database) serializedEnabled() bool {
-	if db == nil {
-		return false
-	}
-	switch db.Driver {
-	case "duckdb", "sqlite", "chai":
-		return true
-	default:
-		return false
-	}
-}
-
-// withSerializedConnection executes the provided function either directly (for engines that
-// tolerate concurrency) or through the pipeline (for single-writer engines). The select/case
-// keeps context cancellation responsive and guarantees that only one job runs at a time when
-// the pipeline is active.
-func (p *serializedPipeline) enqueue(ctx context.Context, job serializedJob) error {
-	lane := p.laneFor(job.kind)
-
-	for {
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		case lane <- job:
-			return nil
-		default:
-			// Yield so other goroutines can place work without blocking; select keeps this non-busy.
-			runtime.Gosched()
-		}
-	}
-}
-
-// laneFor resolves the channel for a workloadKind without a switch so the routing stays
-// channel-centric. If an unknown kind arrives we fall back to the general lane to keep the
-// pipeline robust.
-func (p *serializedPipeline) laneFor(kind WorkloadKind) chan serializedJob {
-	if int(kind) >= 0 && int(kind) < len(p.lanes) {
-		return p.lanes[kind]
-	}
-	return p.lanes[WorkloadGeneral]
-}
-
-// queueFriendlyContext guarantees a minimum timeout so queued work has a fair chance to reach the single
-// worker even while archive imports fill the channel buffers. We avoid mutexes and instead lean on context
-// deadlines to keep the select/case scheduler responsive.
-func queueFriendlyContext(ctx context.Context, min time.Duration) (context.Context, context.CancelFunc) {
-	if min <= 0 {
-		min = serializedWaitFloor
-	}
-	if ctx == nil {
-		return context.WithTimeout(context.Background(), min)
-	}
-	if deadline, ok := ctx.Deadline(); ok {
-		if time.Until(deadline) >= min {
-			return ctx, func() {}
-		}
-		return context.WithTimeout(context.WithoutCancel(ctx), min)
-	}
-	return context.WithTimeout(ctx, min)
-}
-
-// withSerializedConnection queues the given function onto the requested workload lane so
-// long imports or realtime spikes do not block web readers. The default workload keeps
-// backward compatibility while letting specific callers opt into their own lanes.
-func (db *Database) withSerializedConnection(ctx context.Context, fn func(context.Context, *sql.DB) error) error {
-	return db.withSerializedConnectionFor(ctx, WorkloadGeneral, fn)
-}
-
-// withSerializedConnectionFor routes work onto a specific workload queue. We stick to
-// channels/select instead of mutexes so the scheduling remains explicit and follows the
-// Go proverb "Don't communicate by sharing memory; share memory by communicating."
-func (db *Database) withSerializedConnectionFor(ctx context.Context, kind WorkloadKind, fn func(context.Context, *sql.DB) error) error {
-	if db == nil || db.DB == nil {
-		return fmt.Errorf("database unavailable")
-	}
-
-	if !db.serializedEnabled() || db.pipeline == nil {
-		return fn(ctx, db.DB)
-	}
-
-	job := serializedJob{
-		ctx:    ctx,
-		fn:     func(c context.Context, conn *sql.DB) (any, error) { return nil, fn(c, conn) },
-		result: make(chan serializedResult, 1),
-		kind:   kind,
-	}
-
-	if err := db.pipeline.enqueue(ctx, job); err != nil {
-		return err
-	}
-
-	select {
-	case res := <-job.result:
-		return res.err
-	case <-ctx.Done():
-		return ctx.Err()
-	}
-}
-
 // realtimeConverter is configured by the safecastrealtime package to translate
 // raw Safecast units into µSv/h.  Keeping the dependency injected avoids an
 // import cycle and mirrors the Go Proverb "The bigger the interface, the
@@ -392,100 +40,6 @@ var realtimeConverter func(float64, string) (float64, bool)
 // database code can stay agnostic of specific detector logic.
 func SetRealtimeConverter(fn func(float64, string) (float64, bool)) {
 	realtimeConverter = fn
-}
-
-// normalizeDBType trims and lowercases driver names so downstream switch blocks
-// do not miss DuckDB-specific handling just because a caller passed mixed case
-// or incidental whitespace. Centralising the cleanup keeps the checks honest
-// without sprinkling strings.ToLower everywhere. We also map common PostgreSQL
-// aliases onto the pgx driver so placeholder rendering stays consistent and
-// avoids syntax errors like the "?" placeholders seen in server logs.
-func normalizeDBType(dbType string) string {
-	cleaned := strings.ToLower(strings.TrimSpace(dbType))
-
-	switch cleaned {
-	case "postgres", "postgresql", "pq", "postgres+psql", "postgresql+psql":
-		// The project standardises on pgx for PostgreSQL connectivity.
-		// Translating community aliases here keeps SQL builders using
-		// the correct $1 style placeholders instead of "?", preventing
-		// syntax errors during bulk inserts while still honouring the
-		// caller's intent.
-		return "pgx"
-	default:
-		return cleaned
-	}
-}
-
-// startIDGenerator launches a goroutine for generating unique IDs.
-func startIDGenerator(initialID int64) chan int64 {
-	idChannel := make(chan int64)
-	go func(start int64) {
-		currentID := start
-		for {
-			idChannel <- currentID
-			currentID++
-		}
-	}(initialID)
-	return idChannel
-}
-
-// Config holds the configuration details for initializing the database.
-type Config struct {
-	DBType      string // The type of the database driver (e.g., "sqlite", "chai", or "pgx" (PostgreSQL))
-	DBPath      string // The file path to the database file (for file-based databases)
-	DBConn      string // Raw DSN for network drivers (pgx or clickhouse)
-	DBHost      string // The host for PostgreSQL
-	DBPort      int    // The port for PostgreSQL
-	DBUser      string // The user for PostgreSQL
-	DBPass      string // The password for PostgreSQL
-	DBName      string // The name of the PostgreSQL database
-	PGSSLMode   string // The SSL mode for PostgreSQL
-	ClickSecure bool   // Enable TLS when connecting to ClickHouse over HTTP transport
-	Port        int    // The port number (used in database file naming if needed)
-}
-
-// ClickHouseDSNFromConfig assembles a DSN understood by the lightweight HTTP driver.
-// We parse host/port carefully so IPv6 literals keep their brackets intact.
-func ClickHouseDSNFromConfig(cfg Config) string {
-	if trimmed := strings.TrimSpace(cfg.DBConn); trimmed != "" {
-		return trimmed
-	}
-
-	host := strings.TrimSpace(cfg.DBHost)
-	if host == "" {
-		host = "127.0.0.1"
-	}
-
-	if _, _, err := net.SplitHostPort(host); err != nil {
-		port := cfg.DBPort
-		if port <= 0 {
-			port = 9000
-		}
-		host = net.JoinHostPort(host, strconv.Itoa(port))
-	}
-
-	user := strings.TrimSpace(cfg.DBUser)
-	pass := cfg.DBPass
-	name := strings.Trim(strings.TrimSpace(cfg.DBName), "/")
-
-	dsn := url.URL{Scheme: "clickhouse", Host: host}
-	if user != "" {
-		if strings.TrimSpace(pass) != "" {
-			dsn.User = url.UserPassword(user, pass)
-		} else {
-			dsn.User = url.User(user)
-		}
-	}
-	if name != "" {
-		dsn.Path = "/" + name
-	}
-
-	params := url.Values{}
-	if cfg.ClickSecure {
-		params.Set("secure", "true")
-	}
-	dsn.RawQuery = params.Encode()
-	return dsn.String()
 }
 
 // NewDatabase opens DB and configures connection pooling.
